binaries: Add reconstruction scripts under attic/
This is a one-off script that I used to reconstruct the *.buildinfo and
*.changes files for versions 0.0.1 through 0.0.4.

I had uploaded the packages to an aptly apt repo, which deletes and
discards those files.  This script downloads the source packages from
the repo and rebuilds them.  The resulting .deb packages are bit-for-bit
identical to the binary packages from the repo.

The script was run like:

    ./make-debspawn.sh && \
        ./reconstruct-script.sh --sign && \
        debspawn delete reconstruct-config-dlitz-apt-bookworm-amd64